+++ alg/algtest/doc.go
@@ -1,0 +1,2 @@
+// Package algtest provides algorithm test suites.
+package algtest
