+++ alg/binary/binary_test.go
@@ -1,0 +1,11 @@
+package binary
+
+import (
+	"testing"
+
+	"github.com/mmcloughlin/addchain/alg/algtest"
+)
+
+func TestRightToLeft(t *testing.T) {
+	algtest.ChainAlgorithm(t, RightToLeft{})
+}
