+++ internal/test/doc.go
@@ -1,0 +1,2 @@
+// Package test provides testing utilities.
+package test
